yang: Add offline-timer configuration
Add configuration support for the global TACACS+ offline timer:
"system tacplus-options offline-timer <0-3600>".

Changes to the offline-timer configuration are effected by the
vyatta-tacplusd-config script. Using this script requires that the
corresponding configd:end statement is moved to the parent
tacplus-options container. We therefore add a check to the script to
ensure tacplusd is only reloaded when the configuration it cares
about (server container, and offline-timer) changes.

VRVDR-42100