show-tacplus-status.py: Cleanup VRF name handling
Previous logic intended to prevent display of Routing-instance if
TACACS+ is configured in the default routing instance. However the
newline character was not stripped when the name was read from the
env file, so this was never the case.

Clean things up by ensuring the name is stripped of newlines, and
unconditionally display Routing-instance (for screen scraping
backwards compatibility).