Make plugin build flags match default flags
Addresses VRVDR-55985.

Golang plugins need to be built with the paths to source code matching those used when building the module using the plugins, or spurious errors are generated causing the plugin to fail to load due to being 'built with different package versions'. With Go 1.10 / dh-golang 1.38, the trimpath flag was introduced to default (non-overridden) golang builds, and the rules in this repo matched that. However, now that we have upgraded to Go 1.15, we need to handle the additional change to the flag format introduced in Go 1.13 / dh-golang 1.45.

For the record, and for verifying the flags are correct, the dh-golang code looks like this:
```
    if (_go1_has_minor(13)) {
        # Go 1.13 officially supports reproducible build, adding new -trimpath option
        # https://github.com/golang/go/issues/16860
        $this->doit_in_builddir("go", "install", "-trimpath", "-v", @_, @targets);
    } elsif (_go1_has_minor(10)) {
        # Go 1.10 changed flag behaviour, -{gc,asm}flags=all= only works for Go >= 1.10.
        my $trimpath = "all=\"-trimpath=" . $ENV{GOPATH} . "/src\"";
        $this->doit_in_builddir("go", "install", "-gcflags=$trimpath", "-asmflags=$trimpath", "-v", @_, @targets);
    } else {
        $this->doit_in_builddir("go", "install", "-v", @_, @targets);
    }
```
I've verified that this fix works on a vRouter. I've also checked, by tweaking the logic, that each branch of the conditional logic can be hit and that the build flags are set correctly.